(details): Rework /details endpoint
This commit does two major things, it adds a query parameter called type which takes in a string object of the detail they are requesting i.e. either man-duck or ducky as of now. If it is None, then it would send all the details of all types, otherwise it would send the details only of the requested type.

This commit also adds details for man-duck generation and accordingly alters the man-duck generation class